@@ -68,127 +78,132 @@
     "@storybook/react": "^5.3.6",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^9.0.0",
-    "babel-jest": "^24.8.0",
+    "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.2.12",
-    "css-loader": "^3.2.0",
+    "css-loader": "^3.4.2",
     "cssnano": "^4.1.10",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.6.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^4.19.1",
     "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.1.2",
-    "eslint-plugin-react": "^7.12.3",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.18.3",
     "file-loader": "^2.0.0",
-    "jest": "^24.8.0",
-    "mini-css-extract-plugin": "^0.8.0",
-    "node-sass": "^4.11.0",
-    "nodemon": "^1.18.9",
+    "husky": "^4.2.5",
+    "jest": "^24.9.0",
+    "lint-staged": "^10.1.3",
+    "mini-css-extract-plugin": "^0.8.2",
+    "node-sass": "^4.13.1",
+    "nodemon": "^1.19.4",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-cssnext": "^3.1.0",
     "postcss-focus": "^4.0.0",
     "postcss-loader": "^3.0.0",
     "postcss-reporter": "^6.0.1",
-    "react-test-renderer": "^16.6.0",
-    "rimraf": "^2.6.3",
+    "react-test-renderer": "^16.12.0",
+    "rimraf": "^2.7.1",
     "sass-loader": "^6.0.7",
     "storybook-addon-theme-playground": "^1.2.0",
-    "style-loader": "^1.0.0",
-    "terser-webpack-plugin": "^1.4.1",
-    "webpack-cli": "^3.3.7",
-    "webpack-manifest-plugin": "^2.0.4",
+    "style-loader": "^1.1.3",
+    "terser-webpack-plugin": "^1.4.3",
+    "webpack-cli": "^3.3.11",
+    "webpack-manifest-plugin": "^2.2.0",
     "webpack-node-externals": "^1.7.2"
   },
   "engines": {
-    "node": "10.15.0",
-    "npm": "6.4.1"
+    "node": "12.16.1",
+    "npm": "6.13.4"
   },
   "dependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/polyfill": "^7.4.4",
-    "@babel/register": "^7.4.4",
+    "@auth0/s3": "^1.0.0",
+    "@babel/core": "^7.8.4",
+    "@babel/polyfill": "^7.8.3",
+    "@babel/register": "^7.8.3",
     "archiver": "^1.1.0",
-    "async": "^2.6.1",
+    "async": "^2.6.3",
     "axios": "^0.18.1",
     "babel-plugin-styled-components": "^1.10.6",
     "bcrypt-nodejs": "0.0.3",
     "blob-util": "^1.2.1",
     "body-parser": "^1.18.3",
-    "bson-objectid": "^1.2.4",
+    "bson-objectid": "^1.3.0",
     "classnames": "^2.2.6",
     "clipboard": "^1.7.1",
-    "codemirror": "^5.42.2",
+    "codemirror": "^5.52.0",
     "connect-mongo": "^1.3.2",
-    "console-feed": "^2.8.10",
+    "console-feed": "^2.8.11",
     "cookie-parser": "^1.4.3",
     "cors": "^2.8.5",
-    "cross-env": "^5.2.0",
-    "csslint": "^0.10.0",
+    "cross-env": "^5.2.1",
+    "csslint": "^1.0.5",
     "date-fns": "^1.30.1",
     "decomment": "^0.8.7",
     "dotenv": "^2.0.0",
     "dropzone": "^4.3.0",
     "escape-string-regexp": "^1.0.5",
-    "eslint-loader": "^2.1.2",
-    "express": "^4.16.4",
+    "eslint-loader": "^2.2.1",
+    "express": "^4.17.1",
     "express-basic-auth": "^1.2.0",
-    "express-session": "^1.15.6",
-    "friendly-words": "^1.1.3",
+    "express-session": "^1.17.0",
+    "friendly-words": "^1.1.10",
     "htmlhint": "^0.10.1",
     "is-url": "^1.2.4",
-    "jest-express": "^1.10.1",
-    "js-beautify": "^1.8.9",
+    "jest-express": "^1.11.0",
+    "js-beautify": "^1.10.3",
     "jsdom": "^9.8.3",
-    "jshint": "^2.10.2",
+    "jshint": "^2.11.0",
     "lodash": "^4.17.15",
     "loop-protect": "github:catarak/loop-protect",
+    "mime-types": "^2.1.26",
     "mjml": "^3.3.2",
-    "mockingoose": "^2.13.0",
-    "mongoose": "^5.7.5",
+    "mockingoose": "^2.13.2",
+    "mongoose": "^5.9.2",
     "node-uuid": "^1.4.7",
     "nodemailer": "^2.6.4",
     "nodemailer-mailgun-transport": "^1.4.0",
     "passport": "^0.3.2",
     "passport-github": "^1.1.0",
     "passport-google-oauth20": "^1.0.0",
+    "passport-http": "^0.3.0",
     "passport-local": "^1.0.0",
     "pretty-bytes": "^3.0.1",
     "primer-tooltips": "^1.5.11",
     "prop-types": "^15.6.2",
     "q": "^1.4.1",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-helmet": "^5.1.3",
-    "react-hot-loader": "^4.6.3",
+    "react-hot-loader": "^4.12.19",
     "react-inlinesvg": "^0.7.5",
-    "react-redux": "^5.1.1",
-    "react-router": "^3.2.1",
-    "react-split-pane": "^0.1.85",
-    "react-tabs": "^2.3.0",
+    "react-redux": "^5.1.2",
+    "react-router": "^3.2.5",
+    "react-split-pane": "^0.1.89",
+    "react-tabs": "^2.3.1",
     "redux": "^3.7.2",
+    "redux-auth-wrapper": "^2.1.0",
     "redux-devtools": "^3.4.2",
     "redux-devtools-dock-monitor": "^1.1.3",
     "redux-devtools-log-monitor": "^1.4.0",
     "redux-form": "^5.3.3",
     "redux-thunk": "^2.3.0",
-    "request": "^2.88.0",
-    "request-promise": "^4.1.1",
+    "request": "^2.88.2",
+    "request-promise": "^4.2.5",
     "reselect": "^4.0.0",
-    "s3": "^4.4.0",
     "s3-policy": "^0.2.0",
     "sass-extract": "^2.1.0",
     "sass-extract-js": "^0.4.0",
     "sass-extract-loader": "^1.1.0",
-    "shortid": "^2.2.14",
-    "sinon": "^7.3.2",
+    "shortid": "^2.2.15",
+    "sinon": "^7.5.0",
     "sinon-mongoose": "^2.3.0",
-    "slugify": "^1.3.4",
+    "slugify": "^1.3.6",
     "srcdoc-polyfill": "^0.2.0",
     "styled-components": "^5.0.0",
     "styled-theming": "^2.2.0",
     "url": "^0.11.0",
-    "webpack": "^4.39.2",
+    "webpack": "^4.41.6",
     "webpack-dev-middleware": "^2.0.6",
     "webpack-hot-middleware": "^2.24.3",
     "xhr": "^2.5.0"
