@@ -6,39 +6,29 @@ import InlineSVG from 'react-inlinesvg';
 import classNames from 'classnames';
 import * as IDEActions from '../actions/ide';
 import * as FileActions from '../actions/files';
-
-const downArrowUrl = require('../../../images/down-filled-triangle.svg');
-const folderRightUrl = require('../../../images/triangle-arrow-right.svg');
-const folderDownUrl = require('../../../images/triangle-arrow-down.svg');
-const fileUrl = require('../../../images/file.svg');
+import downArrowUrl from '../../../images/down-filled-triangle.svg';
+import folderRightUrl from '../../../images/triangle-arrow-right.svg';
+import folderDownUrl from '../../../images/triangle-arrow-down.svg';
+import fileUrl from '../../../images/file.svg';
 
 export class FileNode extends React.Component {
   constructor(props) {
     super(props);
-    this.renderChild = this.renderChild.bind(this);
-    this.handleKeyPress = this.handleKeyPress.bind(this);
-    this.handleFileNameChange = this.handleFileNameChange.bind(this);
-    this.validateFileName = this.validateFileName.bind(this);
-    this.handleFileClick = this.handleFileClick.bind(this);
-    this.toggleFileOptions = this.toggleFileOptions.bind(this);
-    this.hideFileOptions = this.hideFileOptions.bind(this);
-    this.showEditFileName = this.showEditFileName.bind(this);
-    this.hideEditFileName = this.hideEditFileName.bind(this);
-    this.onBlurComponent = this.onBlurComponent.bind(this);
-    this.onFocusComponent = this.onFocusComponent.bind(this);
 
     this.state = {
       isOptionsOpen: false,
       isEditingName: false,
       isFocused: false,
+      isDeleting: false,
+      updatedName: this.props.name
     };
   }
 
-  onFocusComponent() {
+  onFocusComponent = () => {
     this.setState({ isFocused: true });
   }
 
-  onBlurComponent() {
+  onBlurComponent = () => {
     this.setState({ isFocused: false });
     setTimeout(() => {
       if (!this.state.isFocused) {
@@ -47,41 +37,96 @@ export class FileNode extends React.Component {
     }, 200);
   }
 
-  handleFileClick(e) {
-    e.stopPropagation();
-    if (this.props.name !== 'root' && !this.isDeleting) {
-      this.props.setSelectedFile(this.props.id);
+
+  setUpdatedName = (updatedName) => {
+    this.setState({ updatedName });
+  }
+
+  saveUpdatedFileName = () => {
+    const { updatedName } = this.state;
+    const { name, updateFileName, id } = this.props;
+
+    if (updatedName !== name) {
+      updateFileName(id, updatedName);
     }
   }
 
-  handleFileNameChange(event) {
-    this.props.updateFileName(this.props.id, event.target.value);
+  handleFileClick = (event) => {
+    event.stopPropagation();
+    const { isDeleting } = this.state;
+    const { id, setSelectedFile, name } = this.props;
+    if (name !== 'root' && !isDeleting) {
+      setSelectedFile(id);
+    }
   }
 
-  handleKeyPress(event) {
+  handleFileNameChange = (event) => {
+    const newName = event.target.value;
+    this.setUpdatedName(newName);
+  }
+
+  handleFileNameBlur = () => {
+    this.validateFileName();
+    this.hideEditFileName();
+  }
+
+  handleClickRename = () => {
+    this.setUpdatedName(this.props.name);
+    this.showEditFileName();
+    setTimeout(() => this.fileNameInput.focus(), 0);
+    setTimeout(() => this.hideFileOptions(), 0);
+  }
+
+  handleClickAddFile = () => {
+    this.props.newFile(this.props.id);
+    setTimeout(() => this.hideFileOptions(), 0);
+  }
+
+  handleClickAddFolder = () => {
+    this.props.newFolder(this.props.id);
+    setTimeout(() => this.hideFileOptions(), 0);
+  }
+
+  handleClickUploadFile = () => {
+    this.props.openUploadFileModal(this.props.id);
+    setTimeout(this.hideFileOptions, 0);
+  }
+
+  handleClickDelete = () => {
+    if (window.confirm(`Are you sure you want to delete ${this.props.name}?`)) {
+      this.setState({ isDeleting: true });
+      this.props.resetSelectedFile(this.props.id);
+      setTimeout(() => this.props.deleteFile(this.props.id, this.props.parentId), 100);
+    }
+  }
+
+  handleKeyPress = (event) => {
     if (event.key === 'Enter') {
       this.hideEditFileName();
     }
   }
 
-  validateFileName() {
-    const oldFileExtension = this.originalFileName.match(/\.[0-9a-z]+$/i);
-    const newFileExtension = this.props.name.match(/\.[0-9a-z]+$/i);
-    const hasPeriod = this.props.name.match(/\.+/);
-    const newFileName = this.props.name;
+  validateFileName = () => {
+    const currentName = this.props.name;
+    const { updatedName } = this.state;
+    const oldFileExtension = currentName.match(/\.[0-9a-z]+$/i);
+    const newFileExtension = updatedName.match(/\.[0-9a-z]+$/i);
+    const hasPeriod = updatedName.match(/\.+/);
     const hasNoExtension = oldFileExtension && !newFileExtension;
     const hasExtensionIfFolder = this.props.fileType === 'folder' && hasPeriod;
     const notSameExtension = oldFileExtension && newFileExtension
       && oldFileExtension[0].toLowerCase() !== newFileExtension[0].toLowerCase();
-    const hasEmptyFilename = newFileName === '';
-    const hasOnlyExtension = newFileExtension && newFileName === newFileExtension[0];
+    const hasEmptyFilename = updatedName.trim() === '';
+    const hasOnlyExtension = newFileExtension && updatedName.trim() === newFileExtension[0];
     if (hasEmptyFilename || hasNoExtension || notSameExtension || hasOnlyExtension || hasExtensionIfFolder) {
-      this.props.updateFileName(this.props.id, this.originalFileName);
+      this.setUpdatedName(currentName);
+    } else {
+      this.saveUpdatedFileName();
     }
   }
 
-  toggleFileOptions(e) {
-    e.preventDefault();
+  toggleFileOptions = (event) => {
+    event.preventDefault();
     if (!this.props.canEdit) {
       return;
     }
@@ -93,25 +138,31 @@ export class FileNode extends React.Component {
     }
   }
 
-  hideFileOptions() {
+  hideFileOptions = () => {
     this.setState({ isOptionsOpen: false });
   }
 
-  showEditFileName() {
+  showEditFileName = () => {
     this.setState({ isEditingName: true });
   }
 
-  hideEditFileName() {
+  hideEditFileName = () => {
     this.setState({ isEditingName: false });
   }
 
-  renderChild(childId) {
-    return (
-      <li key={childId}>
-        <ConnectedFileNode id={childId} parentId={this.props.id} canEdit={this.props.canEdit} />
-      </li>
-    );
+  showFolderChildren = () => {
+    this.props.showFolderChildren(this.props.id);
   }
+
+  hideFolderChildren = () => {
+    this.props.hideFolderChildren(this.props.id);
+  }
+
+  renderChild = childId => (
+    <li key={childId}>
+      <ConnectedFileNode id={childId} parentId={this.props.id} canEdit={this.props.canEdit} />
+    </li>
+  )
 
   render() {
     const itemClass = classNames({
@@ -123,150 +174,132 @@ export class FileNode extends React.Component {
       'sidebar__file-item--closed': this.props.isFolderClosed
     });
 
+    const isFile = this.props.fileType === 'file';
+    const isFolder = this.props.fileType === 'folder';
+    const isRoot = this.props.name === 'root';
+
     return (
       <div className={itemClass}>
-        {(() => { // eslint-disable-line
-          if (this.props.name !== 'root') {
-            return (
-              <div className="file-item__content" onContextMenu={this.toggleFileOptions}>
-                <span className="file-item__spacer"></span>
-                {(() => { // eslint-disable-line
-                  if (this.props.fileType === 'file') {
-                    return (
-                      <span className="sidebar__file-item-icon">
-                        <InlineSVG src={fileUrl} />
-                      </span>
-                    );
-                  }
-                  return (
-                    <div className="sidebar__file-item--folder">
-                      <button
-                        className="sidebar__file-item-closed"
-                        onClick={() => this.props.showFolderChildren(this.props.id)}
-                      >
-                        <InlineSVG className="folder-right" src={folderRightUrl} />
-                      </button>
-                      <button
-                        className="sidebar__file-item-open"
-                        onClick={() => this.props.hideFolderChildren(this.props.id)}
-                      >
-                        <InlineSVG className="folder-down" src={folderDownUrl} />
-                      </button>
-                    </div>
-                  );
-                })()}
-                <button className="sidebar__file-item-name" onClick={this.handleFileClick}>{this.props.name}</button>
-                <input
-                  type="text"
-                  className="sidebar__file-item-input"
-                  value={this.props.name}
-                  onChange={this.handleFileNameChange}
-                  ref={(element) => { this.fileNameInput = element; }}
-                  onBlur={() => {
-                    this.validateFileName();
-                    this.hideEditFileName();
-                  }}
-                  onKeyPress={this.handleKeyPress}
-                />
+        { !isRoot &&
+          <div className="file-item__content" onContextMenu={this.toggleFileOptions}>
+            <span className="file-item__spacer"></span>
+            { isFile &&
+              <span className="sidebar__file-item-icon">
+                <InlineSVG src={fileUrl} />
+              </span>
+            }
+            { isFolder &&
+              <div className="sidebar__file-item--folder">
                 <button
-                  className="sidebar__file-item-show-options"
-                  aria-label="view file options"
-                  ref={(element) => { this[`fileOptions-${this.props.id}`] = element; }}
-                  tabIndex="0"
-                  onClick={this.toggleFileOptions}
-                  onBlur={this.onBlurComponent}
-                  onFocus={this.onFocusComponent}
+                  className="sidebar__file-item-closed"
+                  onClick={this.showFolderChildren}
                 >
-                  <InlineSVG src={downArrowUrl} />
+                  <InlineSVG className="folder-right" src={folderRightUrl} />
                 </button>
-                <div className="sidebar__file-item-options">
-                  <ul title="file options">
-                    {(() => { // eslint-disable-line
-                      if (this.props.fileType === 'folder') {
-                        return (
-                          <li>
-                            <button
-                              aria-label="add file"
-                              onClick={() => {
-                                this.props.newFile(this.props.id);
-                                setTimeout(() => this.hideFileOptions(), 0);
-                              }}
-                              onBlur={this.onBlurComponent}
-                              onFocus={this.onFocusComponent}
-                              className="sidebar__file-item-option"
-                            >
-                              Add File
-                            </button>
-                          </li>
-                        );
-                      }
-                    })()}
-                    {(() => { // eslint-disable-line
-                      if (this.props.fileType === 'folder') {
-                        return (
-                          <li>
-                            <button
-                              aria-label="add folder"
-                              onClick={() => {
-                                this.props.newFolder(this.props.id);
-                                setTimeout(() => this.hideFileOptions(), 0);
-                              }}
-                              onBlur={this.onBlurComponent}
-                              onFocus={this.onFocusComponent}
-                              className="sidebar__file-item-option"
-                            >
-                              Add Folder
-                            </button>
-                          </li>
-                        );
-                      }
-                    })()}
-                    <li>
-                      <button
-                        onClick={() => {
-                          this.originalFileName = this.props.name;
-                          this.showEditFileName();
-                          setTimeout(() => this.fileNameInput.focus(), 0);
-                          setTimeout(() => this.hideFileOptions(), 0);
-                        }}
-                        onBlur={this.onBlurComponent}
-                        onFocus={this.onFocusComponent}
-                        className="sidebar__file-item-option"
-                      >
-                        Rename
-                      </button>
-                    </li>
-                    <li>
-                      <button
-                        onClick={() => {
-                          if (window.confirm(`Are you sure you want to delete ${this.props.name}?`)) {
-                            this.isDeleting = true;
-                            this.props.resetSelectedFile(this.props.id);
-                            setTimeout(() => this.props.deleteFile(this.props.id, this.props.parentId), 100);
-                          }
-                        }}
-                        onBlur={this.onBlurComponent}
-                        onFocus={this.onFocusComponent}
-                        className="sidebar__file-item-option"
-                      >
-                        Delete
-                      </button>
-                    </li>
-                  </ul>
-                </div>
+                <button
+                  className="sidebar__file-item-open"
+                  onClick={this.hideFolderChildren}
+                >
+                  <InlineSVG className="folder-down" src={folderDownUrl} />
+                </button>
               </div>
-            );
-          }
-        })()}
-        {(() => { // eslint-disable-line
-          if (this.props.children) {
-            return (
-              <ul className="file-item__children">
-                {this.props.children.map(this.renderChild)}
+            }
+            <button
+              className="sidebar__file-item-name"
+              onClick={this.handleFileClick}
+            >
+              {this.state.updatedName}
+            </button>
+            <input
+              type="text"
+              className="sidebar__file-item-input"
+              value={this.state.updatedName}
+              maxLength="128"
+              onChange={this.handleFileNameChange}
+              ref={(element) => { this.fileNameInput = element; }}
+              onBlur={this.handleFileNameBlur}
+              onKeyPress={this.handleKeyPress}
+            />
+            <button
+              className="sidebar__file-item-show-options"
+              aria-label="view file options"
+              ref={(element) => { this[`fileOptions-${this.props.id}`] = element; }}
+              tabIndex="0"
+              onClick={this.toggleFileOptions}
+              onBlur={this.onBlurComponent}
+              onFocus={this.onFocusComponent}
+            >
+              <InlineSVG src={downArrowUrl} />
+            </button>
+            <div className="sidebar__file-item-options">
+              <ul title="file options">
+                { isFolder &&
+                  <React.Fragment>
+                    <li>
+                      <button
+                        aria-label="add folder"
+                        onClick={this.handleClickAddFolder}
+                        onBlur={this.onBlurComponent}
+                        onFocus={this.onFocusComponent}
+                        className="sidebar__file-item-option"
+                      >
+                        Create folder
+                      </button>
+                    </li>
+                    <li>
+                      <button
+                        aria-label="add file"
+                        onClick={this.handleClickAddFile}
+                        onBlur={this.onBlurComponent}
+                        onFocus={this.onFocusComponent}
+                        className="sidebar__file-item-option"
+                      >
+                        Create file
+                      </button>
+                    </li>
+                    { this.props.authenticated &&
+                      <li>
+                        <button
+                          aria-label="upload file"
+                          onClick={this.handleClickUploadFile}
+                          onBlur={this.onBlurComponent}
+                          onFocus={this.onFocusComponent}
+                        >
+                          Upload file
+                        </button>
+                      </li>
+                    }
+                  </React.Fragment>
+                }
+                <li>
+                  <button
+                    onClick={this.handleClickRename}
+                    onBlur={this.onBlurComponent}
+                    onFocus={this.onFocusComponent}
+                    className="sidebar__file-item-option"
+                  >
+                    Rename
+                  </button>
+                </li>
+                <li>
+                  <button
+                    onClick={this.handleClickDelete}
+                    onBlur={this.onBlurComponent}
+                    onFocus={this.onFocusComponent}
+                    className="sidebar__file-item-option"
+                  >
+                    Delete
+                  </button>
+                </li>
               </ul>
-            );
-          }
-        })()}
+            </div>
+          </div>
+        }
+        { this.props.children &&
+          <ul className="file-item__children">
+            {this.props.children.map(this.renderChild)}
+          </ul>
+        }
       </div>
     );
   }
@@ -288,18 +321,21 @@ FileNode.propTypes = {
   newFolder: PropTypes.func.isRequired,
   showFolderChildren: PropTypes.func.isRequired,
   hideFolderChildren: PropTypes.func.isRequired,
-  canEdit: PropTypes.bool.isRequired
+  canEdit: PropTypes.bool.isRequired,
+  openUploadFileModal: PropTypes.func.isRequired,
+  authenticated: PropTypes.bool.isRequired
 };
 
 FileNode.defaultProps = {
   parentId: '0',
   isSelectedFile: false,
-  isFolderClosed: false
+  isFolderClosed: false,
 };
 
 function mapStateToProps(state, ownProps) {
   // this is a hack, state is updated before ownProps
-  return state.files.find(file => file.id === ownProps.id) || { name: 'test', fileType: 'file' };
+  const fileNode = state.files.find(file => file.id === ownProps.id) || { name: 'test', fileType: 'file' };
+  return Object.assign({}, fileNode, { authenticated: state.user.authenticated });
 }
 
 function mapDispatchToProps(dispatch) {
