@@ -25,7 +25,7 @@ export function serveProject(req, res) {
         const sketchDoc = window.document;
 
         const base = sketchDoc.createElement('base');
-        const fullUrl = `https://${req.get('host')}${req.originalUrl}`;
+        const fullUrl = `${req.protocol}://${req.get('host')}${req.originalUrl}`;
         base.href = `${fullUrl}/`;
         sketchDoc.head.appendChild(base);
 
